backend/requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# ML/AI
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
opencv-python>=4.8.0
pillow>=10.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
